+++ base/error.go
@@ -14,6 +14,7 @@ const noFound = "Not found"
 const accessDenied = "Error 403"
 const resetError = "connection reset by peer"
 const eofError = "unexpected EOF"
+
 //TableFragment table fragment
 const TableFragment = "Table"