+++ e2e/regression/cases/042_cli_transient_schema/bqtail/schema.sql
@@ -1,0 +1,25 @@
+
+
+
+CREATE OR REPLACE TABLE dummy_v${parentIndex} (
+      id         INT64 NOT NULL,
+      type_id    INT64 NOT NULL,
+      name       STRING,
+      segments      ARRAY<STRING>
+);
+
+
+CREATE OR REPLACE TABLE dummy_v${parentIndex}_tmpl (
+        id         INT64 NOT NULL,
+      type_id    INT64 NOT NULL,
+      name       STRING,
+      segments      ARRAY<STRING>
+);
+
+
+CREATE OR REPLACE TABLE dummy_v${parentIndex}_temp_tmpl (
+      id         INT64 NOT NULL,
+      type_id    INT64 NOT NULL,
+      name       STRING,
+      encoded_segments    STRING
+);