+++ tail/sql/builder.go
@@ -3,25 +3,30 @@ package sql
 import (
 	"fmt"
 	"github.com/viant/bqtail/base"
+	"github.com/viant/bqtail/schema"
 	"github.com/viant/bqtail/tail/config"
 	"google.golang.org/api/bigquery/v2"
 	"sort"
 	"strings"
 )
 
-func buildNestedDedupeSQL(sourceTable string, schema Schema, dest *config.Destination) string {
+func buildNestedDedupeSQL(sourceTable string, schema Schema, dest *config.Destination, except map[string]bool) string {
 	projection := make([]string, 0)
-	innerProjection := make([]string, 0)
+	outerProjection := make([]string, 0)
 
 	transform, transformKeys := getTransform(dest)
 	for _, field := range schema.Fields {
 		projection = append(projection, field.Name)
+
 		if expression, ok := getTransformExpression(dest, field); ok {
 			delete(transform, strings.ToLower(field.Name))
-			innerProjection = append(innerProjection, fmt.Sprintf("%v AS %v", expression, field.Name))
+			outerProjection = append(outerProjection, fmt.Sprintf("%v AS %v", expression, field.Name))
 			continue
 		}
-		innerProjection = append(innerProjection, fmt.Sprintf("%v.%v AS %v", dest.Transient.Alias, field.Name, field.Name))
+		if except[field.Name] {
+			continue
+		}
+		outerProjection = append(outerProjection, fmt.Sprintf("%v.%v AS %v", dest.Transient.Alias, field.Name, field.Name))
 	}
 	for _, key := range transformKeys {
 		expression, ok := transform[strings.ToLower(key)]
@@ -29,7 +34,7 @@ func buildNestedDedupeSQL(sourceTable string, schema Schema, dest *config.Destin
 			continue
 		}
 		projection = append(projection, key)
-		innerProjection = append(innerProjection, fmt.Sprintf("%v AS %v", expression, key))
+		outerProjection = append(outerProjection, fmt.Sprintf("%v AS %v", expression, key))
 	}
 
 	return fmt.Sprintf(`SELECT %v
@@ -39,7 +44,7 @@ FROM (
       ROW_NUMBER() OVER (PARTITION BY %v) row_number
   FROM %v $WHERE
 ) %v $JOIN
-WHERE row_number = 1`, strings.Join(innerProjection, ", "), strings.Join(projection, ", "), strings.Join(dest.UniqueColumns, ","), sourceTable, dest.Transient.Alias)
+WHERE row_number = 1`, strings.Join(outerProjection, ", "), strings.Join(projection, ", "), strings.Join(dest.UniqueColumns, ","), sourceTable, dest.Transient.Alias)
 }
 
 func getTransform(dest *config.Destination) (map[string]string, []string) {
@@ -55,15 +60,32 @@ func getTransform(dest *config.Destination) (map[string]string, []string) {
 	return transform, transformKeys
 }
 
-func columnNames(schema *bigquery.TableSchema) []string {
+func columnNames(schema *bigquery.TableSchema, dest *config.Destination, except map[string]bool) []string {
 	var result = make([]string, 0)
+
+	transform, transformKeys := getTransform(dest)
+
 	for _, field := range schema.Fields {
+		if _, ok := transform[strings.ToLower(field.Name)]; ok {
+			delete(transform, strings.ToLower(field.Name))
+		}
+		if except[field.Name] {
+			continue
+		}
 		result = append(result, field.Name)
 	}
+
+	for _, transformKey := range transformKeys {
+		if _, ok := transform[strings.ToLower(transformKey)]; !ok {
+			continue
+		}
+		result = append(result, transformKey)
+	}
+
 	return result
 }
 
-func buildDedupeSQL(sourceTable string, schema Schema, unique map[string]bool, dest *config.Destination) string {
+func buildDedupeSQL(sourceTable string, schema Schema, unique map[string]bool, dest *config.Destination, except map[string]bool) string {
 	var projection = make([]string, 0)
 	var groupBy = make([]string, 0)
 
@@ -79,6 +101,9 @@ func buildDedupeSQL(sourceTable string, schema Schema, unique map[string]bool, d
 		if transformExpression, ok := getTransformExpression(dest, field); ok {
 			delete(transform, strings.ToLower(field.Name))
 			expression = transformExpression
+		}
+		if except[field.Name] {
+			continue
 		}
 		projection = append(projection, fmt.Sprintf("MAX(%v) AS %v", expression, field.Name))
 	}
@@ -102,10 +127,13 @@ GROUP BY %v`,
 	)
 }
 
-func buildSelectAll(sourceTable string, schema Schema, dest *config.Destination) string {
+func buildSelectAll(sourceTable string, schema Schema, dest *config.Destination, except map[string]bool) string {
 	var projection = make([]string, 0)
 	transform, transformKeys := getTransform(dest)
 	for _, field := range schema.Fields {
+		if except[field.Name] {
+			continue
+		}
 		if expression, ok := getTransformExpression(dest, field); ok {
 			delete(transform, strings.ToLower(field.Name))
 			projection = append(projection, fmt.Sprintf("%v AS %v", expression, field.Name))
@@ -137,20 +165,39 @@ func getTransformExpression(dest *config.Destination, field *bigquery.TableField
 	return expression, ok
 }
 
-//BuilAppendDML returns INSERT INTO table () SELECT ...
-func BuilAppendDML(source, destination *bigquery.TableReference, tableScheme *bigquery.TableSchema, dest *config.Destination) string {
-	selectALL := BuildSelect(source, tableScheme, dest)
-	columns := columnNames(tableScheme)
+//BuildAppendDML returns INSERT INTO table () SELECT ...
+func BuildAppendDML(source, destination *bigquery.TableReference, sourceSchema *bigquery.TableSchema, dest *config.Destination, destSchema *bigquery.TableSchema) string {
+	selectALL := BuildSelect(source, sourceSchema, dest, destSchema)
+	except := columnExclusion(sourceSchema, destSchema)
+	columns := columnNames(sourceSchema, dest, except)
 	destTable := base.EncodeTableReference(destination, true)
-	return fmt.Sprintf("INSERT INTO %v(%v) %v", destTable, strings.Join(columns, ","), selectALL)
+	result := fmt.Sprintf("INSERT INTO %v(%v) %v", destTable, strings.Join(columns, ","), selectALL)
+	return result
 }
 
 //BuildSelect returns select SQL statement for specified parameter, if uniqueColumns SQL de-duplicates data
-func BuildSelect(source *bigquery.TableReference, tableScheme *bigquery.TableSchema, dest *config.Destination) string {
-	SQL := buildSelect(source, tableScheme, dest)
+func BuildSelect(source *bigquery.TableReference, sourceSchema *bigquery.TableSchema, dest *config.Destination, destSchema *bigquery.TableSchema) string {
+	except := columnExclusion(sourceSchema, destSchema)
+	SQL := buildSelect(source, sourceSchema, dest, except)
 	join := buildJoins(dest.SideInputs)
 	SQL = strings.Replace(SQL, "$JOIN", join, 1)
 	return SQL
+}
+
+func columnExclusion(source, destSchema *bigquery.TableSchema) map[string]bool {
+	if destSchema == nil {
+		return map[string]bool{}
+	}
+	var excludeMap = make(map[string]bool)
+	sourceMap := schema.IndexFields(source.Fields)
+	destMap := schema.IndexFields(destSchema.Fields)
+	for k := range sourceMap {
+		if _, ok := destMap[k]; !ok {
+			excludeMap[k] = true
+		}
+	}
+	return excludeMap
+
 }
 
 func buildJoins(sideInputs []*config.SideInput) string {
@@ -169,18 +216,18 @@ func buildJoins(sideInputs []*config.SideInput) string {
 }
 
 //buildSelect returns select SQL statement for specified parameter, if uniqueColumns SQL de-duplicates data
-func buildSelect(source *bigquery.TableReference, tableScheme *bigquery.TableSchema, dest *config.Destination) string {
+func buildSelect(source *bigquery.TableReference, tableScheme *bigquery.TableSchema, dest *config.Destination, except map[string]bool) string {
 	sourceTable := "`" + base.EncodeTableReference(source, true) + "`"
 	schema := Schema(*tableScheme)
 	if len(dest.UniqueColumns) == 0 {
-		return buildSelectAll(sourceTable, schema, dest)
+		return buildSelectAll(sourceTable, schema, dest, except)
 	}
 	if schema.IsNested() {
-		return buildNestedDedupeSQL(sourceTable, schema, dest)
+		return buildNestedDedupeSQL(sourceTable, schema, dest, except)
 	}
 	var unique = make(map[string]bool)
 	for _, column := range dest.UniqueColumns {
 		unique[strings.ToLower(column)] = true
 	}
-	return buildDedupeSQL(sourceTable, schema, unique, dest)
+	return buildDedupeSQL(sourceTable, schema, unique, dest, except)
 }